move kit stuff
